Move RequestAnalysis to TBIT
