+++ src/Integration.Vsix.UnitTests/SonarLintTagger/TaggerProviderTests.cs
@@ -30,7 +30,6 @@ using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Tagging;
 using Microsoft.VisualStudio.Utilities;
 using Moq;
-using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Core.Suppression;
 using SonarLint.VisualStudio.Integration.Vsix;
@@ -185,19 +184,6 @@ namespace SonarLint.VisualStudio.Integration.UnitTests
         }
 
         [TestMethod]
-        public void RequestAnalysis_Should_NotThrow_When_AnalysisFails()
-        {
-            mockAnalysisScheduler
-                .Setup(x => x.Schedule("doc1.js", It.IsAny<Action<CancellationToken>>(), It.IsAny<int>()))
-                .Throws<Exception>();
-
-            Action act = () =>
-            provider.RequestAnalysis("doc1.js", "", new[] { AnalysisLanguage.CFamily }, null, null);
-
-            act.Should().NotThrow();
-        }
-
-        [TestMethod]
         [DataRow(null)]
         [DataRow(new string[] { })]
         public void FilterIssueTrackersByPath_NullOrEmptyPaths_AllTrackersReturned(string[] filePaths)
@@ -262,25 +248,6 @@ namespace SonarLint.VisualStudio.Integration.UnitTests
                 });
 
             actual.Should().BeEquivalentTo(trackers);
-        }
-
-        [TestMethod]
-        [DataRow(-1, TaggerProvider.DefaultAnalysisTimeoutMs)]
-        [DataRow(0, TaggerProvider.DefaultAnalysisTimeoutMs)]
-        [DataRow(1, 1)]
-        [DataRow(999, 999)]
-        public void AnalysisTimeout(int envSettingsResponse, int expectedTimeout)
-        {
-            var envSettingsMock = new Mock<IEnvironmentSettings>();
-            envSettingsMock.Setup(x => x.AnalysisTimeoutInMs()).Returns(envSettingsResponse);
-
-            TaggerProvider.GetAnalysisTimeoutInMilliseconds(envSettingsMock.Object).Should().Be(expectedTimeout);
-        }
-
-        [TestMethod]
-        public void AnalysisTimeoutInMilliseconds_NoEnvironmentSettings_DefaultTimeout()
-        {
-            TaggerProvider.GetAnalysisTimeoutInMilliseconds().Should().Be(TaggerProvider.DefaultAnalysisTimeoutMs);
         }
 
         private IIssueTracker[] CreateMockedIssueTrackers(params string[] filePaths) =>
